+++ commands/Fun/say.js
@@ -2,13 +2,14 @@
 module.exports = {
 	name: "say",
 	description: "say it back",
-	guildOnly: true,
 	args: true,
 	aliases: ["repeat"],
 	usage: "<TEXT>",
 	execute(message, args, Discord) {
 		const arg = args.join(" ");
-		message.delete().catch();
+		if(!message.channel.type == "dm"){
+			message.delete().catch();
+		}
 		message.channel.send(arg, {disableEveryone:true});
 	},
 };
